refactor cli to its own module
